Put graphic stroke styling into separate module
